CreateComponent script updated

+++ Build/Tools/PyCodeGenerator/CreateComponentNEW.py
@@ -27,10 +27,10 @@ print ('Library path: '+lib_path.text)
 
 num_args=len(sys.argv)
 
-print('Library creting script started')
+print('Component creting script started')
 
 if num_args <3:
-    exit(0)
+    exit(1)
 
 lib_name=sys.argv[1]
 
@@ -48,16 +48,16 @@ else:
 
 print ('Creating component: '+component_name)
 
-new_path=lib_path.text+namespace_name+'-'+lib_name+'/Core/'
+new_path=lib_path.text+lib_name+'/Core/'
 print (new_path)
 
 if not os.path.exists(new_path) or os.path.isfile(new_path):
     print ('Error access directory!')
-    exit(0)
+    exit(2)
 
 if os.path.exists(new_path+component_name+'.cpp') or os.path.exists(new_path+component_name+'.h'):
     print ('Files already exists!')
-    exit(0)
+    exit(3)
 
 
 # Модифицируем шаблоны библиотеки
@@ -69,13 +69,21 @@ createMapFile(template_data,dest_data,'CodeTemplates/UComponentTemplate.cpp',new
 # открытие файла с компонентой в h-нике библиотеки
 newText = '#include<' +component_name+'.h'+'> \n'
 
-with open(new_path+'U'+lib_name+'.h', 'r') as file:
-    data = file.readlines()
-
-data[6] = newText
-
-with open(new_path+'U'+lib_name+'.h', 'w') as file:
-    file.writelines(data)
+if os.path.exists(new_path+'U'+lib_name+'.h'):
+   with open(new_path+'U'+lib_name+'.h', 'r') as file:
+        data = file.readlines()
+   data.insert(5,newText)
+   with open(new_path+'U'+lib_name+'.h', 'w') as file:
+        file.writelines(data)
+elif os.path.exists(new_path+'Lib.h'):
+   with open(new_path+'Lib.h', 'r') as file:
+        data = file.readlines()
+   data.insert(5,newText)
+   with open(new_path+'Lib.h', 'w') as file:
+        file.writelines(data)
+else:
+    print ('Lib .h file not found!')
+    exit(4)
 
 # изменение cpp-шника библиотеки
 with open('CodeTemplates/ForCPP.txt') as file:
@@ -83,19 +91,48 @@ with open('CodeTemplates/ForCPP.txt') as file:
 text = text + '\n';
 lookup ='//Uppload class here\n'
 
-with open(new_path + 'U' + lib_name + '.cpp', 'r') as file:
-    data = file.readlines()
+if os.path.exists(new_path+'U'+lib_name+'.cpp'):
+   with open(new_path+'U'+lib_name+'.cpp', 'r') as file:
+        data = file.readlines()
 
-with open(new_path + 'U' + lib_name + '.cpp') as file:
-    for num, line in enumerate(file, 1):
-        if lookup in line:
-            writeHere = num
-            data[writeHere] = text
-            break
+   with open(new_path + 'U' + lib_name + '.cpp') as file:
+        for num, line in enumerate(file, 1):
+            if lookup in line:
+               writeHere = num
+               data[writeHere] = text
+               break
 
+   with open(new_path+'U'+lib_name+'.cpp', 'w') as file:
+        file.writelines(data)
+elif os.path.exists(new_path+'Lib.cpp'):
+   with open(new_path+'Lib.cpp', 'r') as file:
+        data = file.readlines()
 
-with open(new_path+'U'+lib_name+'.cpp', 'w') as file:
-    file.writelines(data)
+   with open(new_path + 'Lib.cpp') as file:
+        for num, line in enumerate(file, 1):
+            if lookup in line:
+               writeHere = num
+               data[writeHere] = text
+               break
+
+   with open(new_path+'Lib.cpp', 'w') as file:
+        file.writelines(data)
+else:
+    print ('Lib .cpp file not found!')
+    exit(5)
+
+#with open(new_path + 'U' + lib_name + '.cpp', 'r') as file:
+#    data = file.readlines()
+#
+#with open(new_path + 'U' + lib_name + '.cpp') as file:
+#    for num, line in enumerate(file, 1):
+#        if lookup in line:
+#            writeHere = num
+#            data[writeHere] = text
+#            break
+
+#with open(new_path+'U'+lib_name+'.cpp', 'w') as file:
+#    file.writelines(data)
 
 template_data = ['@CLASS_NAME@', '@SIMPLE_CLASS_NAME@']
 if component_name[0] == 'U':
@@ -104,7 +141,7 @@ if component_name[0] == 'U':
         SimpleComponent_name = SimpleComponent_name[:0] + SimpleComponent_name[(0 + 1):]
         if SimpleComponent_name[0] == 'A':
             SimpleComponent_name = SimpleComponent_name[:0] + SimpleComponent_name[(0 + 1):]
-if component_name[0] == 'R':
+elif component_name[0] == 'R':
     SimpleComponent_name = component_name[:0] + component_name[(0 + 1):]
     if SimpleComponent_name[0] == 'T':
         SimpleComponent_name = SimpleComponent_name[:0] + SimpleComponent_name[(0 + 1):]
@@ -112,26 +149,52 @@ if component_name[0] == 'R':
             SimpleComponent_name = SimpleComponent_name[:0] + SimpleComponent_name[(0 + 1):]
             if SimpleComponent_name[0] == 'A':
                 SimpleComponent_name = SimpleComponent_name[:0] + SimpleComponent_name[(0 + 1):]
+else:
+     SimpleComponent_name = component_name[:0] + component_name[(0 + 1):]
+
 
 
 new_data = [component_name, SimpleComponent_name]
-with open(new_path + 'U' + lib_name + '.cpp') as file:
-    text = file.read()
-    for index, replaced_data in enumerate(template_data):
-        text = text.replace('@UPP' + replaced_data, new_data[index].upper())
-        text = text.replace(replaced_data, new_data[index])
-with open(new_path + 'U' + lib_name + '.cpp', 'w') as file:
-    file.write(text)
+
+if os.path.exists(new_path+'U'+lib_name+'.cpp'):
+   with open(new_path + 'U' + lib_name + '.cpp') as file:
+        text = file.read()
+        for index, replaced_data in enumerate(template_data):
+            text = text.replace('@UPP' + replaced_data, new_data[index].upper())
+            text = text.replace(replaced_data, new_data[index])
+   with open(new_path + 'U' + lib_name + '.cpp', 'w') as file:
+        file.write(text)
+
+elif os.path.exists(new_path+'Lib.cpp'):
+   with open(new_path + 'Lib.cpp') as file:
+        text = file.read()
+        for index, replaced_data in enumerate(template_data):
+            text = text.replace('@UPP' + replaced_data, new_data[index].upper())
+            text = text.replace(replaced_data, new_data[index])
+   with open(new_path + 'Lib.cpp', 'w') as file:
+        file.write(text)
+else:
+    print ('Lib .cpp file not found!')
+    exit(6)
+
+
+#with open(new_path + 'U' + lib_name + '.cpp') as file:
+#    text = file.read()
+#    for index, replaced_data in enumerate(template_data):
+#        text = text.replace('@UPP' + replaced_data, new_data[index].upper())
+#        text = text.replace(replaced_data, new_data[index])
+#with open(new_path + 'U' + lib_name + '.cpp', 'w') as file:
+#    file.write(text)
 
 # добавление h-ника cpp-шника в Qt -> pro файл
 lookupH ='HEADERS +='
 lookupCPP ='SOURCES +='
-with open(lib_path.text+namespace_name+'-'+lib_name+'/Build/Qt/'+namespace_name+'-'+lib_name+'.pro') as file:
+with open(lib_path.text+lib_name+'/Build/Qt/'+lib_name+'.pro') as file:
     data = file.readlines()
 
 textH = '\n    ../../Core/' + component_name+'.h'+' \  \n'
 textCPP = '\n    ../../Core/' + component_name+'.cpp'+' \  \n'
-with open(lib_path.text+namespace_name+'-'+lib_name+'/Build/Qt/'+namespace_name+'-'+lib_name+'.pro') as file:
+with open(lib_path.text+lib_name+'/Build/Qt/'+lib_name+'.pro') as file:
     for num, line in enumerate(file, 1):
         if lookupH in line:
             writeHere = num
@@ -141,39 +204,39 @@ with open(lib_path.text+namespace_name+'-'+lib_name+'/Build/Qt/'+namespace_name+
             data[writeHere] = textCPP
             break
 
-with open(lib_path.text+namespace_name+'-'+lib_name+'/Build/Qt/'+namespace_name+'-'+lib_name+'.pro', 'w') as file:
+with open(lib_path.text+lib_name+'/Build/Qt/'+lib_name+'.pro', 'w') as file:
     file.writelines(data)
 
 #изменение Lib.cpp в Deploy/Include
-lookup ='//#include "../../Core/CPP_FILE_NAME" '
-with open(lib_path.text+namespace_name+'-'+lib_name+'/Deploy/Include/Lib.cpp') as file:
-    data = file.readlines()
+#lookup ='//#include "../../Core/CPP_FILE_NAME" '
+#with open(lib_path.text+lib_name+'/Deploy/Include/Lib.cpp') as file:
+#    data = file.readlines()
 
-text = '\n#include "../../Core/' + component_name+'.cpp"'+'   \n'
-with open(lib_path.text+namespace_name+'-'+lib_name+'/Deploy/Include/Lib.cpp') as file:
-    for num, line in enumerate(file, 1):
-        if lookup in line:
-            writeHere = num
-            data[writeHere] = text
-            break
+#text = '\n#include "../../Core/' + component_name+'.cpp"'+'   \n'
+#with open(lib_path.text+lib_name+'/Deploy/Include/Lib.cpp') as file:
+#    for num, line in enumerate(file, 1):
+#        if lookup in line:
+#            writeHere = num
+#            data[writeHere] = text
+#            break
 
-with open(lib_path.text+namespace_name+'-'+lib_name+'/Deploy/Include/Lib.cpp', 'w') as file:
-    file.writelines(data)
+#with open(lib_path.text+lib_name+'/Deploy/Include/Lib.cpp', 'w') as file:
+#    file.writelines(data)
 
 #обновление заготовок под компилятор CodeBlocks
 lookup ='<Add option="-Wall" />'
-with open(lib_path.text+namespace_name+'-'+lib_name+'/Build/CodeBlocks/'+namespace_name+'-'+ lib_name+'.cbp') as file:
+with open(lib_path.text+lib_name+'/Build/CodeBlocks/'+lib_name+'.cbp') as file:
     data = file.readlines()
 
 text1 = '\n#<Unit filename="../../Core/' + component_name+'.h" />'+'   \n'
 text2 = '#<Unit filename="../../Core/' + component_name+'.cpp" />'+'   \n'
-with open(lib_path.text+namespace_name+'-'+lib_name+'/Build/CodeBlocks/'+namespace_name+'-'+ lib_name+'.cbp') as file:
+with open(lib_path.text+lib_name+'/Build/CodeBlocks/'+lib_name+'.cbp') as file:
     for num, line in enumerate(file, 1):
         if lookup in line:
             writeHere = num
             data[writeHere+1] = text1+text2
             break
 
-with open(lib_path.text+namespace_name+'-'+lib_name+'/Build/CodeBlocks/'+namespace_name+'-'+ lib_name+'.cbp', 'w') as file:
+with open(lib_path.text+lib_name+'/Build/CodeBlocks/'+lib_name+'.cbp', 'w') as file:
     file.writelines(data)
     file.writelines(data)
